Make locations collectable for player's own "Nothing" (#10)
* Update randomizer.py

Adjusted the placement of "Nothing" items so that they still include an item to collect.

Resolves an issue where a Rabi-Ribi player who has their own "Nothing" would not be able to collect that location check, and therefore cannot acquire that location check or hint points in AP. Being able to collect "Nothing" also helps the game behave better with auto-tracking (i.e. Poptracker).

* Added note about "Nothing" item reappearing

The "Nothing" item reappears after being collected. More information probably needed to figure out if this is something that can be fixed, or if it's just Rabi-Ribi's normal game behavior.

* Updated Client for Local "Nothing"

A local "Nothing" item is now treated the same as another player's item in the client when collected. Specifically, it is removed from inventory and deleted from the map. Solves the reappearing problem created by making the item collectable as an exclamation point.

* Remove TODO line